adding otlp python example
