+++ python/Dockerfile.client.lstrace
@@ -1,4 +1,4 @@
 FROM python:3.7
 RUN pip install ls-trace
-ADD client.py /app/client.py
+ADD *.py /app/
 CMD ["ls-trace-run", "/usr/local/bin/python", "/app/client.py"]
