+++ python/Dockerfile.client.otel
@@ -1,16 +1,18 @@
 FROM python:3.7
 RUN pip install flask pymongo redis requests sqlalchemy
 RUN pip install opentelemetry-api \
-    opentelemetry-auto-instrumentation \
+    opentelemetry-instrumentation \
     opentelemetry-sdk \
     opentelemetry-ext-flask \
     opentelemetry-ext-redis \
     opentelemetry-ext-sqlalchemy \
     opentelemetry-ext-requests \
-    opentelemetry-ext-otcollector
+    opentelemetry-ext-otlp \
+    opentelemetry-ext-lightstep
 
 # install instrumentation that has not been released yet
 #RUN pip install \
 #    "git+https://github.com/open-telemetry/opentelemetry-python.git@master#subdirectory=ext/opentelemetry-ext-redis"
-ADD server.py /app/server.py
-CMD ["opentelemetry-auto-instrumentation", "/app/server.py"]
+
+ADD *.py /app/
+CMD ["opentelemetry-instrument", "/app/client.py"]
