+++ python/Dockerfile.server.lstrace
@@ -1,5 +1,5 @@
 FROM python:3.7
 RUN pip install flask redis pymongo sqlalchemy
 RUN pip install ls-trace
-ADD server.py /app/server.py
+ADD *.py /app/
 CMD ["ls-trace-run", "/usr/local/bin/python", "/app/server.py"]
