--- python/client/Dockerfile.noinstrumentation.otel
@@ -1,8 +1,0 @@
-FROM python:3.7
-RUN pip install flask pymongo redis requests sqlalchemy
-RUN pip install opentelemetry-api \
-    opentelemetry-sdk \
-    opentelemetry-ext-otcollector
-
-ADD client.py /app/client.py
-CMD ["/app/client.py"]
